+++ CHANGELOG.md
@@ -9,6 +9,7 @@ HEAD
 * Made sure we don't read more that necessary (issue #422)
 * Fixed error when the key of a `JsonObject` is a `char[]` (issue #423)
 * Reduced code size when using `const` references
+* Fixed error with string of type `unsigned char*` (issue #428)
 
 v5.8.1
 ------